backend/requirements.txt
anyio==3.6.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pocketbase==0.2.2
PyJWT==2.5.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
urllib3==1.26.12
Werkzeug==2.2.2
